move common code to a abstract class
